feat: init clearinghouse extension
