+++ extensions/broker/src/main/java/broker/IdsBrokerServiceImpl.java
@@ -11,11 +11,20 @@
  *       sovity GmbH - initial API and implementation
  *
  */
-package transfer.broker;
+package broker;
 
+import org.eclipse.edc.connector.contract.spi.offer.store.ContractDefinitionStore;
+import org.eclipse.edc.connector.contract.spi.types.offer.ContractDefinition;
 import org.eclipse.edc.protocol.ids.service.ConnectorServiceSettings;
 import org.eclipse.edc.spi.EdcException;
+import org.eclipse.edc.spi.asset.AssetIndex;
+import org.eclipse.edc.spi.event.Event;
+import org.eclipse.edc.spi.event.EventSubscriber;
+import org.eclipse.edc.spi.event.contractdefinition.ContractDefinitionCreated;
+import org.eclipse.edc.spi.event.contractdefinition.ContractDefinitionDeleted;
 import org.eclipse.edc.spi.message.RemoteMessageDispatcherRegistry;
+import org.eclipse.edc.spi.monitor.Monitor;
+import org.eclipse.edc.spi.query.QuerySpec;
 import org.eclipse.edc.spi.system.Hostname;
 import org.eclipse.edc.spi.types.domain.asset.Asset;
 import sender.message.QueryMessage;
@@ -28,11 +37,14 @@ import java.net.URI;
 import java.net.URISyntaxException;
 import java.net.URL;
 import java.util.ArrayList;
+import java.util.HashMap;
+import java.util.HashSet;
 import java.util.List;
+import java.util.Map;
 import java.util.Scanner;
 import java.util.concurrent.ExecutionException;
 
-public class IdsBrokerServiceImpl implements IdsBrokerService {
+public class IdsBrokerServiceImpl implements IdsBrokerService, EventSubscriber {
 
     private static final String CONTEXT_BROKER_REGISTRATION = "BrokerRegistration";
 
@@ -40,13 +52,29 @@ public class IdsBrokerServiceImpl implements IdsBrokerService {
     private final ConnectorServiceSettings connectorServiceSettings;
     private final Hostname hostname;
 
+    private final ContractDefinitionStore contractDefinitionStore;
+
+    private final URL brokerBaseUrl;
+
+    private final AssetIndex assetIndex;
+
+    private final Monitor monitor;
+
     public IdsBrokerServiceImpl(
             RemoteMessageDispatcherRegistry dispatcherRegistry,
             ConnectorServiceSettings connectorServiceSettings,
-            Hostname hostname) {
+            Hostname hostname,
+            ContractDefinitionStore contractDefinitionStore,
+            URL brokerBaseUrl,
+            AssetIndex assetIndex,
+            Monitor monitor) {
         this.dispatcherRegistry = dispatcherRegistry;
         this.connectorServiceSettings = connectorServiceSettings;
         this.hostname = hostname;
+        this.contractDefinitionStore = contractDefinitionStore;
+        this.brokerBaseUrl = brokerBaseUrl;
+        this.assetIndex = assetIndex;
+        this.monitor = monitor;
     }
 
     @Override
@@ -160,5 +188,81 @@ public class IdsBrokerServiceImpl implements IdsBrokerService {
         scanner.close();
 
         return result;
+    }
+
+    @Override
+    public void on(Event<?> event) {
+        if (event instanceof ContractDefinitionCreated contractDefinitionCreated) {
+            //ContractDefinitionCreated -> Send EDC-Asset as IDS-Resource to IDS-Broker
+            handleContractDefinitionCreated(contractDefinitionCreated);
+        } else if (event instanceof ContractDefinitionDeleted contractDefinitionDeleted) {
+            //ContractDefinitionDeleted -> Remove EDC-Asset as IDS-Resource from IDS-Broker
+            handleContractDefinitionDeleted(contractDefinitionDeleted);
+        }
+    }
+
+    private void handleContractDefinitionDeleted(ContractDefinitionDeleted contractDefinitionDeleted) {
+        var eventPayload = contractDefinitionDeleted.getPayload();
+        var contractDefinitionId = eventPayload.getContractDefinitionId();
+        var deletedBrokerIds = getDeletedBrokerIds(brokerBaseUrl, contractDefinitionId);
+
+        removeResourcesFromBroker(deletedBrokerIds, brokerBaseUrl);
+    }
+
+    private void handleContractDefinitionCreated(ContractDefinitionCreated contractDefinitionCreated) {
+        var eventPayload = contractDefinitionCreated.getPayload();
+        var contractDefinitionId = eventPayload.getContractDefinitionId();
+        var contractDefinition = contractDefinitionStore.findById(contractDefinitionId);
+
+        var resourceMap = new HashMap<String, Asset>();
+        var assetsFromContractDefinition = getAssetsFromContractDefinition(contractDefinition);
+        for (var asset : assetsFromContractDefinition) {
+            var resourceId = String.format("%s-%s",
+                    contractDefinition.getId(),
+                    asset.getId());
+            resourceMap.put(resourceId, asset);
+        }
+
+        addResourcesToBroker(resourceMap, brokerBaseUrl);
+    }
+
+    private List<Asset> getAssetsFromContractDefinition(ContractDefinition contractDefinition) {
+        var querySpec = QuerySpec.Builder.newInstance()
+                .filter(new ArrayList<>(contractDefinition.getSelectorExpression().getCriteria()))
+                .build();
+        return assetIndex.queryAssets(querySpec).toList();
+    }
+
+    private List<String> getDeletedBrokerIds(URL brokerBaseUrl, String deletedContractDefinitionIds) {
+        var deletedResourceIds = new ArrayList<String>();
+        var resourceIdsByQuery = findResourceIdsByQuery(brokerBaseUrl, deletedContractDefinitionIds);
+        deletedResourceIds.addAll(resourceIdsByQuery);
+
+        return deletedResourceIds;
+    }
+
+    private void removeResourcesFromBroker(
+            List<String> deletedBrokerResourceIds,
+            URL brokerBaseUrl) {
+        for (var deletedResourceId : deletedBrokerResourceIds) {
+            monitor.info(String.format("Removing resource %s from broker", deletedResourceId));
+            unregisterResourceAtBroker(
+                    brokerBaseUrl,
+                    deletedResourceId);
+        }
+    }
+
+    private void addResourcesToBroker(
+            Map<String, Asset> resourceMap,
+            URL brokerBaseUrl) {
+        var createdResourceSet = new HashSet<>(resourceMap.keySet());
+
+        for (var createdResourceId : createdResourceSet) {
+            monitor.info(String.format("Registering resource %s at broker", createdResourceId));
+            registerResourceAtBroker(
+                    brokerBaseUrl,
+                    createdResourceId,
+                    resourceMap.get(createdResourceId));
+        }
     }
 }
