+++ extensions/clearinghouse/src/main/resources/META-INF/services/org.eclipse.edc.spi.system.ServiceExtension
@@ -1,1 +1,1 @@
-transfer.TransferExtension
+ClearingHouseExtension
