--- extensions/event-provider/src/main/resources/META-INF/services/org.eclipse.edc.spi.system.ServiceExtension
@@ -1,1 +1,0 @@
-org.eclipse.edc.mvd.EventSubscriptionExtension
