+++ settings.gradle.kts
@@ -3,6 +3,4 @@ include("extensions:broker")
 findProject(":extensions:broker")?.name = "broker"
 include("extensions:clearinghouse")
 findProject(":extensions:clearinghouse")?.name = "clearinghouse"
-include("extensions:event-provider")
-findProject(":extensions:event-provider")?.name = "event-provider"
 include("connector")
